Refactored to separate out some components
