+++ webapp/src/components/systemConsole/surveyResults/surveyResults.tsx
@@ -2,11 +2,9 @@
 // See LICENSE.txt for license information.
 
 import client from 'client/client';
-import {format} from 'date-fns';
-import React, {useCallback, useEffect, useMemo, useRef, useState} from 'react';
+import React, {useCallback, useEffect, useMemo, useState} from 'react';
 
-import Button from 'components/common/button/button';
-import {ConfirmationModal} from 'components/systemConsole/surveyResults/confirmationModal';
+import ResultRow from 'components/systemConsole/surveyResults/resultRow';
 
 import type {SurveyResult} from 'types/plugin';
 
@@ -14,9 +12,6 @@ import './style.scss';
 
 function SurveyResults() {
     const [surveyResults, setSurveyResults] = useState<SurveyResult[]>([]);
-    const [showConfirmationModal, setShowConfirmationModal] = useState<boolean>(false);
-    const [surveyReportDownloadingID, setSurveyReportDownloadingID] = useState<string>();
-    const currentSurveyID = useRef<string>();
 
     const hydrateSurveyResults = useCallback((surveyResults: SurveyResult[]) => {
         surveyResults.forEach((surveyResult) => {
@@ -39,20 +34,16 @@ function SurveyResults() {
         fetchSurveyStatus();
     }, [hydrateSurveyResults]);
 
-    const handleStopSurvey = useCallback(async () => {
-        if (!currentSurveyID.current) {
-            return;
-        }
-
+    const handleStopSurvey = useCallback(async (surveyID: string) => {
         try {
-            await client.endSurvey(currentSurveyID.current);
+            await client.endSurvey(surveyID);
         } catch (error) {
             console.error(error);
             return;
         }
 
         const updatedSurveyResult = [...surveyResults];
-        const surveyResultIndex = updatedSurveyResult.findIndex((surveyResult) => surveyResult.id === currentSurveyID.current);
+        const surveyResultIndex = updatedSurveyResult.findIndex((surveyResult) => surveyResult.id === surveyID);
         if (surveyResultIndex < 0) {
             return;
         }
@@ -60,57 +51,6 @@ function SurveyResults() {
         updatedSurveyResult[surveyResultIndex].status = 'ended';
         setSurveyResults(updatedSurveyResult);
     }, [surveyResults]);
-
-    const handleShowConfirmationDialog = useCallback((surveyID: string) => {
-        currentSurveyID.current = surveyID;
-        setShowConfirmationModal(true);
-    }, []);
-
-    const handleHideConfirmationDialog = useCallback(() => {
-        currentSurveyID.current = undefined;
-        setShowConfirmationModal(false);
-    }, []);
-
-    const handleDownloadSurveyReport = useCallback(async (surveyID: string) => {
-        setSurveyReportDownloadingID(surveyID);
-        await client.downloadSurveyReport(surveyID);
-        setSurveyReportDownloadingID('');
-    }, []);
-
-    const generateActions = useCallback((surveyResult: SurveyResult) => {
-        if (surveyResult.status === 'ended') {
-            return (
-                <div
-                    key={surveyResult.id}
-                    className='surveyResultActions'
-                >
-                    <Button
-                        iconClass={surveyReportDownloadingID === surveyResult.id ? 'icon-refresh' : 'icon-download-outline'}
-                        text='Export responses'
-                        onClick={() => handleDownloadSurveyReport(surveyResult.id)}
-                    />
-                </div>
-            );
-        }
-        return (
-            <div
-                key={surveyResult.id}
-                className='surveyResultActions horizontal'
-            >
-                <Button
-                    buttonType='tertiary'
-                    danger={true}
-                    iconClass='icon-flag-checkered'
-                    text='End survey'
-                    onClick={() => handleShowConfirmationDialog(surveyResult.id)}
-                />
-                <Button
-                    iconClass={surveyReportDownloadingID === surveyResult.id ? 'icon-refresh' : 'icon-download-outline'}
-                    onClick={() => handleDownloadSurveyReport(surveyResult.id)}
-                />
-            </div>
-        );
-    }, [handleDownloadSurveyReport, handleShowConfirmationDialog, surveyReportDownloadingID]);
 
     const renderedRows = useMemo(() => {
         if (surveyResults.length === 0) {
@@ -125,25 +65,14 @@ function SurveyResults() {
             );
         }
 
-        return surveyResults.map((result) => {
-            return (
-                <div
-                    key={result.id}
-                    className='horizontal surveyResultRow'
-                >
-                    <div className='startDate'>
-                        <span>{format(result.startDate, 'do MMM yyyy')}</span>
-                        {result.status === 'in_progress' && <div className='badge inProgress'>{'Active'}</div>}
-                    </div>
-                    <div className='endDate'><span>{format(result.endDate, 'do MMM yyyy')}</span></div>
-                    <div className='npsScore'><span>{result.npsScore || '-'}</span></div>
-                    <div className='receiptsCount'><span>{result.receiptCount}</span></div>
-                    <div className='responseCount'><span>{result.responseCount}</span></div>
-                    <div className='actions'>{generateActions(result)}</div>
-                </div>
-            );
-        });
-    }, [generateActions, surveyResults]);
+        return surveyResults.map((result) => (
+            <ResultRow
+                key={result.id}
+                result={result}
+                onStopSurvey={handleStopSurvey}
+            />
+        ));
+    }, [handleStopSurvey, surveyResults]);
 
     return (
         <div className='SurveyResults'>
@@ -165,18 +94,6 @@ function SurveyResults() {
                     {renderedRows}
                 </div>
             </div>
-
-            {
-                showConfirmationModal &&
-                <ConfirmationModal
-                    id='endSurveyConfirmationModal'
-                    title='Are you sure you want to end this survey?'
-                    bodyMessage='Reprehenderit aute laborum anim dolore magna mollit incididunt officia qui labore. Esse culpa ipsum ut esse fugiat do tempor duis mollit. Labore nisi nisi ut aliquip enim aute ex. Esse veniam proident id dolor laborum occaecat voluptate esse laborum cupidatat aliquip ut laborum id. Laborum nostrud officia excepteur. Consectetur ad nisi ullamco est cillum proident Lorem. Aliqua eu est magna deserunt est elit elit do.'
-                    confirmButtonText='End active survey'
-                    handleConfirm={handleStopSurvey}
-                    handleCancel={handleHideConfirmationDialog}
-                />
-            }
 
         </div>
     );
